WIP ClassCamera
