+++ index.ts
@@ -154,6 +154,11 @@ setInterval(() => {
   //q.rotate(new Point(100, 100, 150), Axis.Z, Angle.fromDegrees(5));
 }, 1000 / fps);
 
+
+
+
+
+
 // TODO: method to get rotation of a quad or shapee
 
 // TODO: There needs to be a structure where all instantiated objects are added to a superArray,
@@ -189,138 +194,21 @@ setInterval(() => {
 //      only explicit method getSomething() and setSomething().
 
 // TODO: replace Positional interface with this
-interface PositionalFinal {
-  getRotation(Point, Axis): Angle;
-  setRotation(Point, Axis, Angle): PositionalFinal;
-  rotate(Point, Axis, Angle): PositionalFinal;
-
-  getPosition(): Point;
-  setPosition(Point): PositionalFinal;
-  translate(Vector): PositionalFinal;
-}
-
-interface DrawableFinal {
-  strokeColor: Color;
-  fillColor: Color;
-  setColor(color: Color);
-  getStrokeColor(): Color;
-  setStrokeColor(strokeColor: Color);
-  getFillColor(): Color;
-  setFillColor(fillColor: Color);
-
-  draw(canvas: Canvas);
-}
-
-type RenderableFinal = PositionalFinal & DrawableFinal;
-
-class PositionalObject implements PositionalFinal {
-  constructor(public x, public y, public z) {}
-
-  _getAxisDiff(pivot: Point, axis: Axis) {
-    let dx = 0,
-      dy = 0;
-    if (axis == Axis.X) {
-      dx = this.y - pivot.y;
-      dy = this.z - pivot.z;
-    } else if (axis == Axis.Y) {
-      dx = this.x - pivot.x;
-      dy = this.z - pivot.z;
-    } else if (axis == Axis.Z) {
-      dx = this.x - pivot.x;
-      dy = this.y - pivot.y;
-    }
-    return { dx, dy };
-  }
-
-  getRotation(pivot: Point, axis: Axis): Angle {
-    let diff = this._getAxisDiff(pivot, axis);
-    let currentAngle = new Angle(Math.atan2(diff.dy, diff.dx));
-    return currentAngle;
-  }
-
-  setRotation(pivot: Point, axis: Axis, angle: Angle): PositionalFinal {
-    let diff = this._getAxisDiff(pivot, axis);
-    let radius = Math.sqrt(diff.dx ** 2 + diff.dy ** 2);
-    let newX = pivot.x + radius * Math.cos(angle.radians);
-    let newY = pivot.x + radius * Math.sin(angle.radians);
-    if (axis == Axis.X) {
-      this.y = newX;
-      this.z = newY;
-    } else if (axis == Axis.Y) {
-      this.x = newX;
-      this.z = newY;
-    } else if (axis == Axis.Z) {
-      this.x = newX;
-      this.y = newY;
-    }
-    return this;
-  }
-
-  rotate(pivot: Point, axis: Axis, angle: Angle): PositionalFinal {
-    let currentAngle = this.getRotation(pivot, axis);
-    this.setRotation(pivot, axis, currentAngle.add(angle));
-    return this;
-  }
-
-  getPosition(): Point {
-    return new Point(this.x, this.y, this.z);
-  }
-
-  setPosition(point: Point): PositionalFinal {
-    this.x = point.x;
-    this.y = point.y;
-    this.z = point.z;
-    return this;
-  }
-
-  translate(vec: Vector): PositionalFinal {
-    this.x += vec.x;
-    this.y += vec.y;
-    this.z += vec.z;
-    return this;
-  }
-}
-
-class Camera extends PositionalObject {
-  public frustum: Frustum;
-  constructor(
-    public width,
-    public height,
-    public horizontalFOV = Angle.fromDegrees(45),
-    public verticalFOV = horizontalFOV
-  ) {
-    super(0, 0, 0);
-    this.frustum = new Frustum(width, height, horizontalFOV, verticalFOV);
-  }
-
-  toCameraSpace(point: Point): Point {
-    // TODO
-    return point;
-  }
-
-  fromCameraSpaceToScreenSpace(point: Point): Point {
-    return point;
-  }
-
-  toScreenSpace(point: Point) {
-    return this.fromCameraSpaceToScreenSpace(this.toCameraSpace(point));
-  }
-}
 
 // Renderable encapsulator
 // keeps a reference to a Renderable object to
 // calculate and store the camera space position.
 // The storage is important for optimization.
-class CameraSpaceRenderable {
-  constructor(public object: Renderable, public camera: Camera) {}
-}
+// class CameraSpaceRenderable {
+//   constructor(public object: Renderable, public camera: Camera) {}
+// }
 
-class ScreenSpaceRenderable {
-  public frustum: Frustum;
-  constructor(public object: CameraSpaceRenderable, public canvas: Canvas) {
-    this.frustum = object.camera.frustum;
-  }
-}
+// class ScreenSpaceRenderable {
+//   public frustum: Frustum;
+//   constructor(public object: CameraSpaceRenderable, public canvas: Canvas) {
+//     this.frustum = object.camera.frustum;
+//   }
+// }
 
 /**
  * The structure of the rendering pipeline:
